pyproject.toml
[tool.poetry]
name = "ai_dungeon"
version = "0.1.0"
description = ""
authors = ["Alexander Batyrgariev <sasha@bashkort.org>"]

[tool.poetry.dependencies]
python = "^3.8"
python-decouple = "^3.3"
aioredis = "^1.3.1"
gql = {version = "^3.0.0-alpha.1", allow-prereleases = true}
kutana = "^4.2.0"
httpx = "^0.13.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
